Agendamento Controller

+++ src/Backend/MinhaAgendaDeConsultas.Domain/Repositorios/Agendamento/IAgendamentoConsultasConsultasOnlyRepositorio.cs
@@ -11,7 +11,7 @@ namespace MinhaAgendaDeConsultas.Domain.Repositorios
         /// <param name="pacienteId"></param>
         /// <param name="medicoId"></param>
         /// <returns>Retorna uma lista de Agendamentos Disponíveis</returns>
-        public Task<bool> GetDisponibilides(DateTime DataDeInicio, DateTime DataFim, int? medicoId = null);
+        public Task<bool> GetDisponibilides(DateTime DataDeInicio, DateTime DataFim, long? medicoId = null);
         public Task<IList<AgendamentoConsultas>> GetAgendamentosMedico(int medicoId);
         public Task<IList<AgendamentoConsultas>> GetAgendamentosPaciente(int pacienteId);
     }
